--- src/main/java/com/amuzil/omegasource/magus/skill/conditionals/enums/ActivationType.java
@@ -1,12 +1,0 @@
-package com.amuzil.omegasource.magus.skill.conditionals.enums;
-
-/**
- * 4 main types of activation. Each of these have sub-types. Again, I'll define them as an enum for now and
- * consolidate for good code practice later.
- */
-public enum ActivationType {
-    MENU,
-    HOTKEY,
-    MULTIKEY,
-    GESTURE
-}
