--- src/main/java/com/amuzil/omegasource/magus/skill/conditionals/enums/ExecuteType.java
@@ -1,6 +1,0 @@
-package com.amuzil.omegasource.magus.skill.conditionals.enums;
-
-public enum ExecuteType {
-    DESIGNATED,
-    SPECIAL
-}
