--- src/main/java/com/amuzil/omegasource/magus/skill/conditionals/enums/GestureType.java
@@ -1,10 +1,0 @@
-package com.amuzil.omegasource.magus.skill.conditionals.enums;
-
-/**
- * Covers a range of VR gestures for those who aren't too flexible or athletically inclined, and those
- * who are quite adept in their range of motion.
- */
-public enum GestureType {
-    SIMPLE,
-    COMPLEX
-}
