--- src/main/java/com/amuzil/omegasource/magus/skill/conditionals/enums/InitialiseType.java
@@ -1,10 +1,0 @@
-package com.amuzil.omegasource.magus.skill.conditionals.enums;
-
-/**
- * These 2 types are generally used for multikey. Initialise is how to start listening for the hotkey or key combo,
- * and the execution is when to stop.
- */
-public enum InitialiseType {
-    TOGGLE,
-    HOLD
-}
