--- src/main/java/com/amuzil/omegasource/magus/skill/conditionals/enums/RadialType.java
@@ -1,6 +1,0 @@
-package com.amuzil.omegasource.magus.skill.conditionals.enums;
-
-public enum RadialType {
-    MOUSE,
-    NUMBER
-}
